implemented accessors, removed setCoef, and more.
I implemented getCoefficients and getFormat. getCoefficients only works
if the matrix is assembled. I removed the setCoef from the ELL class
because if any getCoefficients get changed or added, the matrix should
probably be re-assembled anyways. Finally, I started to make a function
to find the row with the most non-zero values, but im having trouble
finding out how to iterate through _buildCoeff