updated matvec tests and made some interface fixes
I give sparseMatrix a matvec interface because my tests wouldn't work
unless I did. I think this is because i made it a sparseMatrix type instead
of sparseMatrix_CSR type. what is the proper type to use?

Added a matvec test that gives empty input and a normal input for a
3x3 sparse matrix. the normal input one won't pass right now because i did the assertion wrong. I'll have to fix that later.

In any case, the matvec method did nto compile for me.