templating on matvec test with other fixes
many fixes were made with the matvec tests, the templating feature was
added, and it was added back to the cmake list. I modified both the
header and class file so that assemble is a method and the code still
compiles. However, a lot of work is need on the assemble method still.