Changed getCoefficient to operator() and more
I changed the way the getCoefficient operator was defined. I fixed the
tests implemented by others so that they would run with the new framework.
I implemented the assemble method so the tests would pass. This could merge
with main but should probably be integrated with the other ell branch before